+++ src/cdc_XR21B1411.cpp
@@ -28,7 +28,7 @@ uint32_t XR21B1411::Init(uint32_t parent, uint32_t port, uint32_t lowspeed) {
         uint8_t buf[constBufSize];
         USB_DEVICE_DESCRIPTOR * udd = reinterpret_cast<USB_DEVICE_DESCRIPTOR*>(buf);
 
-        uint8_t rcode;
+        uint32_t rcode;
         UsbDeviceDefinition *p = NULL;
         EpInfo *oldep_ptr = NULL;
         uint8_t num_of_conf; // number of configurations
