+++ src/cdcacm.cpp
@@ -51,7 +51,7 @@ uint32_t ACM::Init(uint32_t parent, uint32_t port, uint32_t lowspeed) {
         uint8_t buf[constBufSize];
         USB_DEVICE_DESCRIPTOR * udd = reinterpret_cast<USB_DEVICE_DESCRIPTOR*>(buf);
 
-        uint8_t rcode;
+        uint32_t rcode;
         UsbDeviceDefinition *p = NULL;
         EpInfo *oldep_ptr = NULL;
         uint8_t num_of_conf; // number of configurations
@@ -271,7 +271,7 @@ uint32_t ACM::Release() {
 }
 
 uint32_t ACM::Poll() {
-        //uint8_t rcode = 0;
+        //uint32_t rcode = 0;
         //if(!bPollEnable)
         //        return 0;
         //return rcode;
